Add rpc pubsub metrics with minor bugfix
